front end

+++ routes/ideas.js
@@ -44,19 +44,31 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updateIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    if (idea.username === req.body.username) {
+      const updateIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      {
-        new: true,
-      }
-    );
-    res.json({ success: true, data: updateIdea });
+        {
+          new: true,
+        }
+      );
+      return res.json({ success: true, data: updateIdea });
+    }
+
+    // No match
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: 'You are not authorized to update this resource',
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -71,8 +83,18 @@ router.put('/:id', async (req, res) => {
 // Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorised to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     res.json({ success: false, error: 'Something went wrong' });
